requirements.txt
python-dotenv
pandas
numpy
matplotlib
pymongo
certifi
pymongo[srv]==4.7
scikit-learn
dill
pyaml
xgboost
mlflow
# -e .